+++ src/keybinds/files_tab.rs
@@ -2,7 +2,8 @@ use std::str::FromStr;
 
 use ratatui::crossterm::event::KeyEvent;
 
-use super::HelpSection;
+use super::HelpItem;
+use super::Section;
 use super::Shortcut;
 use super::config::FilesTabKeybindsConfig;
 use super::keybinds_store::KeybindsStore;
@@ -60,15 +61,12 @@ impl FilesTabKeybinds {
             .unwrap_or(FilesTabEvent::Unbound)
     }
 
-    pub fn make_help(&self) -> Vec<HelpSection> {
-        vec![HelpSection::new(
-            "Main panel",
-            make_keybinds_help!(
-                self.keys,
-                FilesTabEvent::Untrack => "untrack file",
-                FilesTabEvent::Restore => "restore file",
-            ),
-        )]
+    pub fn make_help(&self) -> Vec<HelpItem> {
+        make_keybinds_help!(
+            self.keys,
+            FilesTabEvent::Untrack => Section::Files, "untrack file",
+            FilesTabEvent::Restore => Section::Files, "restore file",
+        )
     }
 }