+++ src/ui/evolog_tab.rs
@@ -25,7 +25,7 @@ use crate::keybinds::DetailsPanelEvent;
 use crate::keybinds::DetailsPanelKeybinds;
 use crate::keybinds::EvologTabEvent;
 use crate::keybinds::EvologTabKeybinds;
-use crate::keybinds::HelpSection;
+use crate::keybinds::HelpItem;
 use crate::ui::AppAction;
 use crate::ui::Component;
 use crate::ui::ComponentInputResult;
@@ -195,11 +195,11 @@ impl Tab for EvologTab<'_> {
         Ok(self.context_menu(self.entry_panel.selected_position()))
     }
 
-    fn make_main_panel_help(&self) -> Vec<HelpSection> {
+    fn make_main_panel_help(&self) -> Vec<HelpItem> {
         self.keybinds.make_help()
     }
 
-    fn make_details_panel_help(&self) -> Vec<(String, String)> {
+    fn make_details_panel_help(&self) -> Vec<HelpItem> {
         self.details_keybinds.make_help()
     }
 }